+++ koni_gzip/test/decompressed_source_test.dart
@@ -139,7 +139,7 @@ final class _HeadFormat extends ArchiveFormat {
   }
 
   @override
-  Future<ArchiveReader> openReader(
+  Future<ArchiveReader> createReader(
     ByteSource source,
     ArchiveReadOptions options,
   ) async => _FixedReader(this, source);